refactor: move reporter
