+++ packages/log/ts/reporter.ts
@@ -1,6 +1,18 @@
 import { config } from './config.js'
+import { createConsoleLogReporter } from './console/index.js'
+import { createColorLogReporter } from './platform/index.js'
 import { store } from './store.js'
-import { LogEntry } from './types.js'
+import type { LogEntry } from './types.js'
+
+export function getDefaultReporter() {
+  try {
+    return createColorLogReporter()
+  }
+  catch (e) {
+    // istanbul ignore next
+    return createConsoleLogReporter()
+  }
+}
 
 export function writeToReporters(logEntry: LogEntry, filter: (reporterId: string) => boolean) {
   writeToReporters.fn(logEntry, filter)
