--- packages/log/ts/writeToReporters.spec.ts
@@ -1,9 +1,0 @@
-import delay from 'delay'
-import { getLogger } from './getLogger.js'
-
-test('auto config in first write', async () => {
-  const log = getLogger('config-in-first-write')
-  log.emergency('first: should see this message')
-  log.emergency('second: should see this message')
-  await delay(10)
-})
